Creation of 4LearningHtml
This, as described in the title as well as the document itself, is meant as a guide to help those who desire to learn the basics of Html. This is in no means a perfect guide and I do appreciate any feedback as to changes that could be applied to this document.